Create mod.ts to publish on deno add ability to save from id add console.logs

+++ src/commands/save.ts
@@ -6,23 +6,35 @@ type Directory = {
 	};
 };
 
+// deno-lint-ignore no-explicit-any
 function get_files(project: any) {
-	let retval: Directory = { directory: {} };
-	for (let file_name in project) {
+	const retval: Directory = { directory: {} };
+	for (const file_name in project) {
 		const file = project[file_name];
 		if ('directory' in file) {
 			retval.directory[file_name] = get_files(file.directory);
 		} else {
-			retval.directory[file_name] = {
-				file: { contents: new Uint8Array(file.file.contents.buffer) },
-			};
+			// handle very old repls with files that are still strings
+			if (typeof file.file.contents !== 'string') {
+				retval.directory[file_name] = {
+					file: {
+						contents: new Uint8Array(file.file.contents.buffer),
+					},
+				};
+			} else {
+				retval.directory[file_name] = {
+					file: {
+						contents: new TextEncoder().encode(file.file.contents),
+					},
+				};
+			}
 		}
 	}
 	return retval;
 }
 
 async function write_files(files: Directory, base_path: string) {
-	for (let file_name in files.directory) {
+	for (const file_name in files.directory) {
 		const file = files.directory[file_name];
 		if ('directory' in file) {
 			try {
@@ -39,35 +51,80 @@ async function write_files(files: Directory, base_path: string) {
 			}
 			write_files(file, `${base_path}/${file_name}`);
 		} else {
-			await Deno.writeFile(
-				`${base_path}/${file_name}`,
-				file.file.contents,
-				{
-					create: true,
-				},
-			);
+			try {
+				await Deno.writeFile(
+					`${base_path}/${file_name}`,
+					file.file.contents,
+					{
+						create: true,
+					},
+				);
+				console.log(
+					`Writing "${base_path}/${file_name}" to the disk...`,
+				);
+			} catch (_e) {
+				console.error(
+					`Failed to write "${base_path}/${file_name}" to the disk...`,
+				);
+			}
 		}
 	}
 }
 
 export async function save(url: string, destination: string) {
-	const actual_url = new URL(`${url}.json`);
+	let actual_url;
+	try {
+		actual_url = new URL(`${url}.json`);
+	} catch (_e) {
+		actual_url = new URL(`https://sveltelab.dev/${url}.json`);
+	}
 	const is_correct_url = [
 		'sveltelab.dev',
 		'www.sveltelab.dev',
 		'localhost',
 	].includes(actual_url.hostname);
 	if (!is_correct_url) {
-		throw new Error('Invalid URL: it must be a sveltelab project');
+		console.log('Invalid URL: it must be a sveltelab project');
+		return;
 	}
-	const project = await fetch(actual_url).then((res) => res.json());
 	try {
-		await Deno.mkdir(destination, { recursive: true });
-	} catch (_e) {
-		await Deno.remove(destination, { recursive: true });
-		await Deno.mkdir(destination, { recursive: true });
-	}
-	const files = get_files(project.files);
+		console.log('Fetching the REPL...');
+		const project = await fetch(actual_url).then((res) => {
+			if (!res.ok) throw new Error();
+			return res.json();
+		});
+		let already_made_directory = false;
+		try {
+			const dir = await Deno.stat(destination);
+			if (dir.isFile || dir.isDirectory) {
+				const want_to_delete = confirm(
+					`The destination ${
+						dir.isFile
+							? "exist and it's a file"
+							: 'folder already exist'
+					}, do you want to delete it?`,
+				);
+				if (want_to_delete) {
+					await Deno.remove(destination, { recursive: true });
+					await Deno.mkdir(destination, { recursive: true });
+					already_made_directory = true;
+				} else if (dir.isFile) {
+					console.error('Abort saving...');
+					return;
+				}
+			}
+		} catch (_e) {}
+		if (!already_made_directory) {
+			try {
+				await Deno.mkdir(destination, { recursive: true });
+			} catch (_e) {
+				/* empty */
+			}
+		}
+		const files = get_files(project.files);
 
-	await write_files(files, destination);
+		await write_files(files, destination);
+	} catch (_e) {
+		console.error("Can't fetch the REPL, please try again.");
+	}
 }
